+++ NEWS.md
@@ -2,7 +2,7 @@
 
 * Added function *is_country()* to test whether a string is a country name or related to list of countries names
 * Added function *is_date()* to test whether values are dates
-* Added function *is_key()* to test whether a set of columns could be the keys of a table
+* Added function *is_keycol()* to test whether a set of columns could be the keys of a table
 * Added function *find_countrycol()* to automate search of columns containing country names in data frames
 * Added function *find_timecol()* to automate search of date and/or year columns in data frames
 * Speed improvements for the functions *country_name()* and *match_table()*
@@ -14,7 +14,7 @@
 
 * Function *country_name()* for converting and translating country names based on a fuzzy matching approach
 * Function *match_table()* to easily construct conversion tables with the help of fuzzy matching algorithm
-* Function *get_mode()* returns the mode of a vectors (most frequent observation)
+* Function *mode()* returns the mode of a vectors (most frequent observation)
 * Dataset *country_reference_list* containing a country names in different languages and naming conventions
 * Added vignettes and documentation
 * Created website for the package